Update SQL_bootcamp_challenges.sql
NOTE: here the rows which do not have a match in the right table will be returned and the part of the second table that is empty will be filled with NULL place holders. Also if we want, in the query, we can place the table names in different lines, one above another and the result will be the same.